Implemented Initial Sync Engine

+++ fixture_setup/src/source_models/user_stats.rs
@@ -1,5 +1,5 @@
 #[derive(Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct UserStatsSource {
-    pub last_played: Option<i32>,
+    pub last_played: Option<i64>,
 }
